update other agent
