+++ assets/js/base/context/cart-checkout/checkout/index.js
@@ -3,8 +3,10 @@
  */
 import { PluginArea } from '@wordpress/plugins';
 import { CURRENT_USER_IS_ADMIN } from '@woocommerce/block-settings';
-import BlockErrorBoundary from '@woocommerce/base-components/block-error-boundary';
-import { SlotFillProvider } from '@woocommerce/blocks-checkout';
+import {
+	SlotFillProvider,
+	BlockErrorBoundary,
+} from '@woocommerce/blocks-checkout';
 import { getSetting } from '@woocommerce/settings';
 
 /**
