+++ assets/js/base/utils/render-frontend.js
@@ -2,7 +2,7 @@
  * External dependencies
  */
 import { render, Suspense } from '@wordpress/element';
-import BlockErrorBoundary from '@woocommerce/base-components/block-error-boundary';
+import { BlockErrorBoundary } from '@woocommerce/blocks-checkout';
 
 // Some blocks take care of rendering their inner blocks automatically. For
 // example, the empty cart. In those cases, we don't want to trigger the render
