+++ assets/js/blocks/cart-checkout/checkout/frontend.js
@@ -11,7 +11,7 @@ import {
 	StoreNoticesProvider,
 	ValidationContextProvider,
 } from '@woocommerce/base-context';
-import BlockErrorBoundary from '@woocommerce/base-components/block-error-boundary';
+import { BlockErrorBoundary } from '@woocommerce/blocks-checkout';
 import { CURRENT_USER_IS_ADMIN } from '@woocommerce/block-settings';
 import { createInterpolateElement } from 'wordpress-element';
 import {
