+++ assets/js/blocks/single-product/edit/index.js
@@ -5,7 +5,7 @@ import { __ } from '@wordpress/i18n';
 import { useState } from '@wordpress/element';
 import { Placeholder, Button, PanelBody } from '@wordpress/components';
 import { withProduct } from '@woocommerce/block-hocs';
-import BlockErrorBoundary from '@woocommerce/base-components/block-error-boundary';
+import { BlockErrorBoundary } from '@woocommerce/blocks-checkout';
 import EditProductLink from '@woocommerce/editor-components/edit-product-link';
 import { singleProductBlockPreview } from '@woocommerce/resource-previews';
 import { InspectorControls } from '@wordpress/block-editor';
