+++ packages/checkout/index.js
@@ -7,3 +7,4 @@ export { default as ExperimentalOrderMeta } from './order-meta';
 export { default as ExperimentalOrderShippingPackages } from './order-shipping-packages';
 export { default as Panel } from './panel';
 export { SlotFillProvider } from 'wordpress-components';
+export { default as BlockErrorBoundary } from './error-boundary';
